Clean up development XCM config too
